add ACL and auto assign done

+++ app/Http/Controllers/Admin/ConferenceManager/AnnouncementsController.php
@@ -10,34 +10,36 @@ use App\Models\Announcements;
 class AnnouncementsController extends BaseConferenceController
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @param $conferenceId
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index($conferenceId)
     {
+        $this->authorize('view-announcement');
         $announcements = Announcements::where('conference_id', $conferenceId)->get();
         return view('layouts.admin.conference_manager.announcements.list', compact('announcements'));
     }
 
     /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
+     * @param $conferenceId
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create($conferenceId)
     {
+        $this->authorize('create-announcement');
         return view('layouts.admin.conference_manager.announcements.create', ['conference_id' => $conferenceId]);
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
+        $this->authorize('create-announcement');
     	$request->validate([
     		'title' => ['required', 'string', 'max:255'],
             'expiry_date' => ['required', 'date'],
@@ -61,26 +63,28 @@ class AnnouncementsController extends BaseConferenceController
     }
 
     /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param $conferenceId
+     * @param $id
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit($conferenceId, $id)
     {
+        $this->authorize('update-announcement');
         $announcement = Announcements::find($id);
         return view('layouts.admin.conference_manager.announcements.edit', ['conference_id' => $this->conferenceId, 'announcement'=>$announcement]);
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param $conferenceId
+     * @param $id
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, $conferenceId, $id)
     {
+        $this->authorize('update-announcement');
     	$request->validate([
     		'title' => ['required', 'string', 'max:255'],
             'expiry_date' => ['required', 'date'],
@@ -102,13 +106,14 @@ class AnnouncementsController extends BaseConferenceController
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param $conferenceId
+     * @param $id
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($conferenceId, $id)
     {
+        $this->authorize('delete-announcement');
         $announcement = Announcements::find($id);
 
         if ($announcement->delete()) {
